refactor: рефакторинг страницы Movies по замечаниям
