+++ src/pages/Movies/index.js
@@ -3,7 +3,7 @@ import CardFilm from '../../components/Cards/CardFilm/CardFilm';
 import CardAnnotation from '../../components/Cards/CardAnnotation/CardAnnotation';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
 import TagsList from '../../components/TagsList/TagsList';
-import { TitleH1, PseudoButtonTag, Loader } from '../../components/utils/index';
+import { Loader, PseudoButtonTag, TitleH1 } from '../../components/utils';
 
 export {
   BasePage,
