+++ src/pages/Movies/locales/RU.js
@@ -1,4 +1,4 @@
-import { MAIN_PAGE_TITLE, MOVIES_TITLE } from '../config/routes';
+import { MAIN_PAGE_TITLE, MOVIES_TITLE } from '../../../config/routes';
 
 const moviesPageTexts = {
   headTitle: `${MAIN_PAGE_TITLE} | ${MOVIES_TITLE}`,
